refactor(debug_server): extract /video and /call families to debug_server_call (refs #332)
Wave 23b · 10th per-family extract.  All 12 video/call control
handlers move verbatim from debug_server.c into a new sibling
module debug_server_call.{c,h}, mirroring the pattern of the
prior 9 family extracts (#340-#348).

Endpoints (all bearer-auth):
  POST /video/start         POST /video/stop          GET  /video
  POST /video/show          POST /video/hide
  POST /video/call/start    POST /video/call/end
  GET  /call/status         POST /call/mute           POST /call/unmute
  POST /call/minimize       POST /call/restore

Helpers moved alongside their handlers:
  - show_pane_async, hide_pane_async, minimize_pane_async, restore_pane_async
    (LVGL-thread hops via tab5_lv_async_call)

Same convention as the prior family extracts: `check_auth` /
`send_json_resp` mapped via local #define to the
`tab5_debug_check_auth` / `_send_json_resp` shared helpers in
debug_server_internal.h.  Identical behaviour.

debug_server.c LOC: 3,339 → ~3,090 (-249).
debug_server_call.c: 0 → 305.

E2E verification on physical Tab5 192.168.1.90:
  - story_smoke 14/14 in 66s (matches baseline)
  - POST /video/call/start  → {"ok":true,"fps":5}
  - GET  /call/status       → in_call=true, pane_visible=true,
                              uplink_active=true, voice_mode=CALL,
                              full heap snapshot included
  - GET  /video             → fps=5, frames_sent=8 after 5s
  - POST /video/call/end    → {"ok":true}
